Error statistics

This function takes a fitted WRTDS model and computes error statistics the residuals used here are cross-validation residuals, which will be slightly larger than regular regression residuals in the case of censored data, the residuals are computed from random residuals computed from makeAugmentedSample(), the function returns a list of error statistics and also prints them to the console

Usage

errorStats(eList)

Arguments

eList

named list with at least the Daily, Sample, and INFO dataframes

Value

erStats a numeric vector consisting of the following statistics RsqLC the R squared value for predictions of ln(Concentration) RsqLF the R squared value for predictions of ln(Flux) rmse the root mean squared error for ln(Concentration), same value would apply for Flux sepPercent the standard error of prediction for Concentration, expressed in percent same value would apply for Flux

Examples

eList <- Choptank_eList
errorStats(eList)
#> 
#>  Root Mean Squared Error in natural log units =  0.295
#>  Rsquared for natural log of concentration    =  0.459
#>  Rsquared for natural log of flux             =  0.95
#>  Standard error of estimate = 30.1 %
#>   RsqLogC RsqLogF  rmse sepPercent
#> 1   0.459    0.95 0.295       30.1