Estimates all daily values of Concentration, Flux, Flow-Normalized Concentration, and Flow Normalized Flux

Uses the surfaces estimated in estSurfaces to estimate these four time series in addition to the time series for standard error and yHat (estimated log concentration). The results are stored in an augmented version of the Daily data frame, which is returned as part of an EGRET object.

Bin the LogQ values by day-of-year.

estDailyFromSurfaces(eList, localsurfaces = NA, localDaily = NA)

getConcFluxFromSurface(eList, allLogQsByDayOfYear, localDaily,
  localsurfaces = NA)

getSurfaceEstimates(eList, localsurfaces = NA, localDaily = NA)

bin_Qs(localDaily)

Arguments

eList
named list with at least the Daily and INFO dataframes, and the surface matrix

localsurfaces
surface over-riding the one stored in eList. Default is NA.

localDaily
data frame to override eList$Daily. Default is NA.

allLogQsByDayOfYear
list

Value

egret object with altered Daily dataframe

Daily dataframe with yHat, SE, ConcDay and FluxDay calulated

Examples

eList <- Choptank_eList
#################################################
# This is usually done in modelEstimation:
Daily <- getDaily(eList)
surfaceIndexParameters<-surfaceIndex(Daily)
INFO <- eList$INFO
INFO$bottomLogQ<-surfaceIndexParameters[['bottomLogQ']]
INFO$stepLogQ<-surfaceIndexParameters[['stepLogQ']]
INFO$nVectorLogQ<-surfaceIndexParameters[['nVectorLogQ']]
INFO$bottomYear<-surfaceIndexParameters[['bottomYear']]
INFO$stepYear<-surfaceIndexParameters[['stepYear']]
INFO$nVectorYear<-surfaceIndexParameters[['nVectorYear']]
eList$INFO <- INFO
#################################################

Daily <- estDailyFromSurfaces(eList)