Computes several values of the flow duration curve for streamflow centered on a specific date of the year

This function is useful for helping the analyst determine the empirical probability distribution of streamflow for a particular part of the year or for the whole year. This is particularly useful in setting up discharge scales for various other plots in this package.

flowDuration(eList, centerDate = "09-30", qUnit = 2, span = 365)

Arguments

eList
named list with at least Daily and INFO dataframes

centerDate
character specifying the center date of the part of the year for which the flow duration is to be calculated, it is in the form "mm-dd" (it must be in quotes). Default is "09-30"

qUnit
object of qUnit class printqUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name. Default is qUnit = 2, which corresponds to cubic meters per second.

span
number this is the half-width of the window over which the discharge values are to be used in constructing the flow-duration curve. If the full year is desired any value greater than 182 will provide serve. Note that for a window of about 2-months width, a span value should be about 30. Default is 365.

Value

qDuration A named vector with flow duration information.

Examples

eList <- Choptank_eList
# for a window of 30 days either side of June 25 expressed in units of cfs:
flowDuration(eList,"06-25", qUnit=1,span=30)

#> 
#> Flow Duration for Choptank River 
#> 
#> Flow duration period is centered on June 25 
#> And spans the period from May 26  To July 25
#> 
#> Discharge units are Cubic Feet per Second 

#>    min     5%    10%    25%    50%    75%    90%    95%    max 
#>    3.6   13.0   16.0   30.0   52.0  100.0  213.0  335.0 3830.0 

# for a flow-duration curve covering the whole year, expressed in units of csf: 
flowDuration(eList, "01-01", qUnit=2)

#> 
#> Flow Duration for Choptank River 
#> 
#> Flow duration is based on full year
#> 
#> Discharge units are Cubic Meters per Second 

#>          min           5%          10%          25%          50%          75% 
#> 9.910896e-03 3.398022e-01 4.530695e-01 9.344559e-01 2.406932e+00 4.615646e+00 
#>          90%          95%          max 
#> 8.211885e+00 1.308238e+01 2.463566e+02