Produces annual series of 8 streamflow statistics (and a lowess smooth of them) from daily streamflow data

Part of the flowHistory system. The data come from Daily and INFO data frames. Note that the function setPA must be run before this to establish the period of analysis (e.g. water year).

makeAnnualSeries(eList, edgeAdjust = TRUE)

Arguments

eList
named list with at least Daily and INFO dataframes

edgeAdjust
logical specifying whether to use the modified method for calculating the windows at the edge of the record. The modified method tends to reduce curvature near the start and end of record. Default is TRUE, but a logical in INFO$edgeAdjust will override the default.

Value

annualSeries matrix that contains the annual series of streamflow statistics annualSeries is a matrix 3 * 8 * numYears, where numYears is the number of years in the data set in the first dimension 1 is the year, 2 is the actual value, 3 is the smoothed value in the second dimension, the index is the istat value (identifying the flow statistic) the third dimension is year

Details

istat	Name

1	minimum 1-day daily mean discharge

2	minimum 7-day mean of the daily mean discharges

3	minimum 30-day mean of the daily mean discharges

4	median of the daily mean discharges

5	mean of the daily mean discharges

6	maximum 30-day mean of the daily mean discharges

7	maximum 7-day mean of the daily mean discharges

8	maximum 1-day daily mean discharge

The smooth is a loess smooth computed on the log flow values and then transformed back to real space Smoothing window is a fixed number of years, the window width has a default value of 20 years but can be modified by changing its value in INFO data frame (using setPA function)

Examples

eList <- Choptank_eList
annualSeries <- makeAnnualSeries(eList)