mergeReport

This function does three things. 1) It transfers the daily discharge value from the Daily data frame to to Sample data frame for those days with samples. 2) It merges the INFO, Daily and Sample data frames to form an eList object, 3) and it prints out a "report" of basic information about the Daily and Sample data frames.

mergeReport(INFO, Daily, Sample = NA, surfaces = NA, verbose = TRUE,
  interactive = NULL)

Arguments

INFO
dataframe metadata about the Sample and Daily data frames.

Daily
dataframe containing the daily discharge data

Sample
dataframe containing the sample data

surfaces
matrix returned from modelEstimation. Default is NA.

verbose
logical specifying whether or not to display summary information on the Daily and Sample dataframes.

interactive
logical deprecated. Use 'verbose' instead

Value

eList named list with Daily, Sample, and INFO dataframes, along with the surfaces matrix. Any of these values can be NA, not all EGRET functions will work with missing parts of the named list eList.

Details

Note that the Sample dataframe in the global environment does not update with the flow information.

See also

Examples

siteNumber <- '01491000'
pCode <- '00631'
# NOT RUN {
Daily <- readNWISDaily(siteNumber,'00060', '1984-10-01', '')
Sample <- readNWISSample(siteNumber,pCode, '1984-10-01', '')
INFO <- readNWISInfo(siteNumber,pCode,interactive=FALSE)
eList <- mergeReport(INFO, Daily, Sample)
Sample <- eList$Sample
plot(eList)

# Create eList with no water quality data:

eList <- mergeReport(INFO, Daily, Sample = NA)
plotFour(eList)
# }