Print annual results for a given streamflow statistic

Part of the flowHistory system. The index of the flow statistics is istat. These statistics are: (1) 1-day minimum, (2) 7-day minimum, (3) 30-day minimum, (4) median (5) mean, (6) 30-day maximum, (7) 7-day maximum, and (8) 1-day maximum.

printSeries(eList, istat, qUnit = 1, runoff = FALSE)

Arguments

eList
named list with at least the Daily and INFO dataframes

istat
A numeric value for the flow statistic to be graphed (possible values are 1 through 8)

qUnit
object of qUnit class printqUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name. Default is 1, which is cubic feet per second.

runoff
logical variable, if TRUE the streamflow data are converted to runoff values in mm/day

Value

data frame with:

years	integer	year

qActual	numeric	Actual flow statistic (based on istat)

qSmooth	numeric	Smoothed flow statistic

years	integer

Examples

eList <- Choptank_eList
printReturn <- printSeries(eList, 5)

#> 
#>  Choptank River
#>  Water Year
#>     mean daily
#>     Cubic Feet per Second
#>    year   annual   smoothed
#>            value    value
#> 
#>    1980    150.2      108
#>    1981     78.3      109
#>    1982    107.6      111
#>    1983    176.1      112
#>    1984    201.9      114
#>    1985     53.6      115
#>    1986     92.8      117
#>    1987    119.1      118
#>    1988     66.0      120
#>    1989    198.2      121
#>    1990    141.5      123
#>    1991     97.0      124
#>    1992     77.2      126
#>    1993    131.8      127
#>    1994    193.7      129
#>    1995     85.3      131
#>    1996    220.5      132
#>    1997    206.0      134
#>    1998    172.5      136
#>    1999    102.3      138
#>    2000    166.8      139
#>    2001    172.2      141
#>    2002     43.8      143
#>    2003    305.2      145
#>    2004    186.4      147
#>    2005    134.6      149
#>    2006    126.8      151
#>    2007    151.2      153
#>    2008     90.5      155
#>    2009    130.0      157
#>    2010    254.0      159
#>    2011    185.2      161