Reminder to user of flow Unit properties such as cubic meters per second or thousands of cubic feet per second.

Cheat sheet to print out pre-defined qUnit properties from qUnit class. Flow units included:

Number	ObjectName	shortName

unitFactor		1	cfs

Cubic Feet per Second	35.31467		2

cms	Cubic Meters per Second	1

3	thousandCfs	Thousand Cubic Feet per Second	0.03531467

4	thousandCms	Thousand Cubic Meters per Second

0.001		5	mmDay

mm per day			6

mmYear	mm per year

printqUnitCheatSheet()

Examples

printqUnitCheatSheet()

#> The following codes apply to the qUnit list:
#> 1 =  cfs  ( Cubic Feet per Second )
#> 2 =  cms  ( Cubic Meters per Second )
#> 3 =  thousandCfs  ( Thousand Cubic Feet per Second )
#> 4 =  thousandCms  ( Thousand Cubic Meters per Second )