Processing of Water Quality Data

Processes water quality data. This function looks at detection limit and detection conditions to determine if a value is left censored or not. Censored values are given the qualifier "<". The dataframe is also converted from a long to wide format.

processQWData(data, pCode = TRUE)

Arguments

data
dataframe from Water Quality Portal

pCode
logical if TRUE, assume data came from a pCode search, if FALSE, characteristic name.

Value

data dataframe with first column dateTime, and at least one qualifier and value columns (subsequent qualifier/value columns could follow depending on the number of parameter codes)

See also

Examples

# NOT RUN {
library(dataRetrieval)

rawWQP <- readWQPqw('21FLEECO_WQX-IMPRGR80','Phosphorus', '', '')
Sample2 <- processQWData(rawWQP, pCode=FALSE)
# }