Import NWIS Sample Data for EGRET analysis

Imports data from NWIS web service. This function gets the data from here: https://nwis.waterdata.usgs.gov/nwis/qwdata/ A list of parameter and statistic codes can be found here: https://help.waterdata.usgs.gov/codes-and-parameters For raw data, use readNWISqw from the dataRetrieval package. This function will retrieve the raw data, and compress it (summing constituents). See section 3.2.4 of the vignette for more details.

readNWISSample(siteNumber, parameterCd, startDate = "", endDate = "",
  verbose = TRUE, interactive = NULL)

Arguments

siteNumber
character USGS site number. This is usually an 8 digit number

parameterCd
character USGS parameter code. This is usually an 5 digit number.

startDate
character starting date for data retrieval in the form YYYY-MM-DD.

endDate
character ending date for data retrieval in the form YYYY-MM-DD.

verbose
logical specifying whether or not to display progress message

interactive
logical deprecated. Use 'verbose' instead

Value

A data frame 'Sample' with the following columns:

Name	Type	Description

Date	Date	Date

ConcLow	numeric	Lower limit of concentration

ConcHigh	numeric	Upper limit of concentration

Uncen	integer	Uncensored data (1=TRUE, 0=FALSE)

ConcAve	numeric	Average concentration

Julian	integer	Number of days since Jan. 1, 1850

Month	integer	Month of the year [1-12]

Day	integer	Day of the year [1-366]

DecYear	numeric	Decimal year

MonthSeq	integer	Number of months since January 1, 1850

SinDY	numeric	Sine of the DecYear

CosDY	numeric	Cosine of the DecYear

See also

Examples

# NOT RUN {
# These examples require an internet connection to run

Sample_01075 <- readNWISSample('01594440','01075', '1985-01-01', '1985-03-31')
# }