Import user-supplied sample data for EGRET analysis

Imports data from a user-supplied file, and converts it to a Sample data frame (including summing multiple constituents), appropriate for EGRET analysis. First column is date, second is remark code, and third is value. If multiple constituents are to be combined with interval censoring, additional columns can be inserted, each starting with remark code (specifically looking for <), and values.

readUserSample(filePath, fileName, hasHeader = TRUE, separator = ",",
  verbose = TRUE, interactive = NULL)

Arguments

filePath
character specifying the path to the file

fileName
character name of file to open

hasHeader
logical true if the first row of data is the column headers

separator
character character that separates data cells. , default is "," which is separator used in a .csv file.

verbose
logical specifying whether or not to display progress message

interactive
logical deprecated. Use 'verbose' instead

Value

A data frame 'Sample' with the following columns:

Name	Type

Description		Date

Date	Date

ConcLow	numeric	Lower limit of concentration

ConcHigh	numeric

Upper limit of concentration		Uncen

integer	Uncensored data (1=TRUE, 0=FALSE)

ConcAve	numeric	Average concentration

Julian	integer

Number of days since Jan. 1, 1850		Month

integer	Month of the year [1-12]

Day	integer	Day of the year [1-366]

DecYear	numeric

Decimal year		MonthSeq

integer	Number of months since January 1, 1850

SinDY	numeric	Sine of the DecYear

CosDY	numeric

Cosine of the DecYear		Name

See also

Examples

filePath <- system.file("extdata", package="EGRET")
fileName <- 'ChoptankRiverNitrate.csv'
Sample <- readUserSample(filePath,fileName, separator=";",verbose=FALSE)