Surface date limits

Sets the Date limits to the surfaces being estimated from the Sample data set. The start is less than a year prior to the first date (typically the date of the first sample) and the end is less than a year after the last date (typically the date of the last sample). The start is constrained to be on the first day of the period of analysis and the end is constrained to be on the last day of the the period of analysis

surfaceStartEnd(paStart, paLong, Date1, Date2)

Arguments

paStart
numeric integer specifying the starting month for the period of analysis, 1<=paStart<=12, default is 10

paLong
numeric integer specifying the length of the period of analysis, in months, 1<=paLong<=12, default is 12

Date1
Date set to Date of earliest data in Sample.

Date2
Date set to Date of latest data in Sample.

Examples

eList <- Choptank_eList
Date1 <- eList$Sample$Date[1]
Date2 <- range(eList$Sample$Date)[2]
surfaceStartEnd(10, 12, Date1, Date2)

#> $surfaceStart
#> [1] "1979-10-01"
#> 
#> $surfaceEnd
#> [1] "2011-09-30"
#>