EGRET package includes WRTDS and flowHistory

Package:

EGRET

Type:	Package

License:

Unlimited for this package, dependencies have more restrictive licensing.

Copyright:	This software is in the public domain because it contains materials

that originally came from the United States Geological Survey, an agency of	the United States Department of Interior. For more information, see the

official USGS copyright policy at	https://www.usgs.gov/visual-id/credit_usgs.html#copyright

LazyLoad:

yes

Details

Collection of functions to do WRTDS and flowHistory analysis, and produce graphs and tables of data and results from these analyses.

References

Hirsch, R.M., and De Cicco, L.A., 2014, User guide to Exploration and Graphics for RivEr Trends (EGRET) and dataRetrieval: R packages for hydrologic data: U.S. Geological Survey Techniques and Methods book 4, chap. A10, 94 p., https://doi.org/10.3133/tm4A10