Import metadata to create INFO data frame

Populates INFO data frame from either NWIS (readNWISInfo), Water Quality Portal (readWQPInfo), or user-supplied files (readUserInfo).

readNWISInfo(siteNumber, parameterCd, interactive = TRUE)

readWQPInfo(siteNumber, parameterCd, interactive = TRUE)

readUserInfo(filePath, fileName, hasHeader = TRUE, separator = ",",
  interactive = TRUE)

Arguments

siteNumber
character site number. For readNWISInfo, this is usually an 8 digit number, for readWQPInfo, it is usually a longer code. For instance, a USGS site number in the Water Quality Portal would be in the form `USGS-XXXXXXXX`. If the siteNumber is left blank (an empty string), the interactive option allows users to enter required information by hand, otherwise those fields are left blank.

parameterCd
character USGS parameter code (a 5 digit number) or characteristic name (if using readWQPInfo). If the parameterCd is left blank (an empty string), the interactive option allows users to enter required information by hand, otherwise those fields are left blank.

interactive
logical Option for interactive mode. If true, there is user interaction for error handling and data checks.

filePath
character specifying the path to the file (used in readUserInfo)

fileName
character name of file to open (used in readUserInfo)

hasHeader
logical true if the first row of data is the column headers (used in readUserInfo)

separator
character that separates data cells (used in readUserInfo)

Value

INFO data frame. Any metadata can be stored in INFO. However, there are 8 columns that EGRET uses by name in some functions:

Required column	Used in function	Description

param.units***	All concentration plotting functions	The units as listed in this field are used to create the concentration axis labels

shortName	All plotting functions	Station short name, used to label plots

paramShortName	All plotting functions	Parameter short name, used to label plots

drainSqKm	plotFlowSingle, printSeries	Calculate runoff

constitAbbrev	saveResults	Parameter abbrieviation, used to auto-name workspace

staAbbrev	saveResults	Station abbrieviation, used to auto-name workspace

paStart	Most EGRET functions	Starting month of period of analysis. Defaults to 10

paLong	Most EGRET functions	Length in number of months of period of analysis. Defaults to 12

*** Additionally, EGRET assumes that all concentrations are saved in mg/l. If some variation of 'mg/l' is not found in INFO$param.units, functions that calculate flux will issue a warning. This is because the conversion from mg/l to the user-specified flux unit (e.g., kg/day) uses hard-coded conversion factors.

See also

Examples

# These examples require an internet connection to run
# Automatically gets information about site 05114000 and temperature
# \donttest{
INFO <- readNWISInfo('05114000','00010')

#> Your site for streamflow data is:
#>  05114000 .
#> Your site name is SOURIS RIVER NR SHERWOOD, ND 
#> but you can modify this to a short name in a style you prefer. 
#> This name will be used to label graphs and tables. 
#> If you want the program to use the name given above, just do a carriage return,
#> otherwise enter the preferred short name(no quotes):
#> 
#> The latitude and longitude of the site are:  48.99002 ,  -101.9582 (degrees north and west).
#> 
#> The drainage area at this site is  8940 square miles
#>  which is being stored as 23154.49 square kilometers.
#> 
#> It is helpful to set up a station abbreviation when doing multi-site studies,
#> enter a unique id (three or four characters should work). It is case sensitive.  
#> Even if you don't feel you need an abbreviation for your site you need to enter something(no quotes):
#> 
#> Your water quality data are for parameter number:
#> 00010 
#> which has the name:' Temperature, water, degrees Celsius '.
#> Typically you will want a shorter name to be used in graphs and tables.
#> The suggested short name is:' Temperature, water '.
#> If you would like to change the short name, enter it here, 
#> otherwise just hit enter (no quotes):
#> The units for the water quality data are:  deg C .
#> It is helpful to set up a constiuent abbreviation, enter a unique id 
#> three or four characters should work something like tn or tp or NO3).
#> Even if you don't feel you need an abbreviation you need to enter something (no quotes):
#> 
#> Required concentration units are mg/l. 
#> The INFO dataframe indicates: deg C 
#> Flux calculations will be wrong if units are not consistent.

# }
# These examples require an internet connection to run
# Automatically gets information about site 01594440 and temperature, no interaction with user
nameToUse <- 'Specific conductance'
pcodeToUse <- '00095'
# \donttest{
INFO <- readWQPInfo('USGS-04024315',pcodeToUse)

#> Your site for data is USGS-04024315 .
#> Your site name is LITTLE BALSAM CREEK NEAR PATZAU, WI ,
#> but you can modify this to a short name in a style you prefer. 
#> This name will be used to label graphs and tables. 
#> If you want the program to use the name given above, just do a carriage return, 
#> otherwise enter the preferred short name(no quotes):
#> 
#> Your water quality data are for parameter number 00095 
#> which has the name:' Specific conductance, water, unfiltered, microsiemens per centimeter at 25 degrees Celsius '.
#> Typically you will want a shorter name to be used in graphs and tables. 
#> The suggested short name is:' Specific conductance '.
#> If you would like to change the short name, enter it here, otherwise just hit enter (no quotes):
#> It is helpful to set up a constiuent abbreviation when doing multi-constituent studies,
#> enter a unique id (three or four characters should work something like tn or tp or NO3).
#> Even if you don't feel you need an abbreviation you need to enter something (no quotes):
#> 
#> It is helpful to set up a station abbreviation when doing multi-site studies, 
#> enter a unique id (three or four characters should work).
#> Even if you don't feel you need an abbreviation for your site you need to enter something(no quotes):
#> 
#> Required concentration units are mg/l. 
#> The INFO dataframe indicates: uS/cm @25C 
#> Flux calculations will be wrong if units are not consistent

INFO2 <- readWQPInfo('WIDNR_WQX-10032762',nameToUse)

#> Your site for data is WIDNR_WQX-10032762 .
#> Your site name is North Branch Pheasant Branch downstream of pond at Deming Way ,
#> but you can modify this to a short name in a style you prefer. 
#> This name will be used to label graphs and tables. 
#> If you want the program to use the name given above, just do a carriage return, 
#> otherwise enter the preferred short name(no quotes):
#> 
#> Your water quality data are for parameter number  
#> which has the name:' Specific conductance '.
#> Typically you will want a shorter name to be used in graphs and tables. 
#> The suggested short name is:' Specific conductance '.
#> If you would like to change the short name, enter it here, otherwise just hit enter (no quotes):
#> Water Quality Portal does not offer a simple method to obtain unit information.
#> EGRET requires concentration units in mg/l for all functions to work correctly. 
#> Enter the concentration units of Sample data:
#> 
#> It is helpful to set up a constiuent abbreviation when doing multi-constituent studies,
#> enter a unique id (three or four characters should work something like tn or tp or NO3).
#> Even if you don't feel you need an abbreviation you need to enter something (no quotes):
#> 
#> It is helpful to set up a station abbreviation when doing multi-site studies, 
#> enter a unique id (three or four characters should work).
#> Even if you don't feel you need an abbreviation for your site you need to enter something(no quotes):
#> 
#> No drainage area (and/or units) was listed in the WQP site file for this site.
#> Drainage area is used to calculate runoff parameters in flow history calculations.
#> Please enter the drainage area, you can enter it in the units of your choice.
#> Enter the area, then enter drainage area code, 
#> 1 is square miles, 
#> 2 is square kilometers, 
#> 3 is acres, 
#> 4 is hectares.
#> Area(no quotes):
#> 
#> Unit Code (1-4, no quotes):
#> Required concentration units are mg/l. 
#> The INFO dataframe indicates:  
#> Flux calculations will be wrong if units are not consistent

# To adjust the label names:
INFO$shortName <- "Little"
INFO$paramShortName <- "SC"
# }
filePath <- system.file("extdata", package="EGRET")
fileName <- 'infoTest.csv'
INFO <- readUserInfo(filePath,fileName, separator=",",interactive=FALSE)