Import NWIS Daily Data for EGRET analysis

Imports daily data from NWIS web service. This function gets the data from here: https://waterservices.usgs.gov/

readNWISDaily(
  siteNumber,
  parameterCd = "00060",
  startDate = "",
  endDate = "",
  verbose = TRUE,
  interactive = NULL,
  convert = TRUE
)

Arguments

siteNumber
character USGS site number. This is usually an 8 digit number

parameterCd
character USGS parameter code. This is usually an 5 digit number.

startDate
character starting date for data retrieval in the form YYYY-MM-DD.

endDate
character ending date for data retrieval in the form YYYY-MM-DD.

verbose
logical specifying whether or not to display progress message

interactive
logical deprecated. Use 'verbose' instead

convert
logical Option to include a conversion from cfs to cms (35.314667). The default is TRUE, which is appropriate for using NWIS data in the EGRET package. Set this to FALSE to not include the conversion. If the parameter code is not 00060 (NWIS discharge), there is no conversion applied.

Value

A data frame 'Daily' with the following columns:

Name	Type	Description

Date	Date	Date

Q	numeric	Discharge in m^3/s

Julian	integer	Number of days since Jan. 1, 1850

Month	integer	Month of the year [1-12]

Day	integer	Day of the year [1-366]

DecYear	numeric	Decimal year

MonthSeq	integer	Number of months since January 1, 1850

Qualifier	character	Qualifying code

i	integer	Index of days, starting with 1

LogQ	numeric	Natural logarithm of Q

Q7	numeric	7 day running average of Q

Q30	numeric	30 day running average of Q

See also

Examples

# \donttest{

Daily <- readNWISDaily('01594440','00060', '1985-01-01', '1985-03-31')

#> There are 90 data points, and 90 days.

DailySuspSediment <- readNWISDaily('01594440','80154', '1985-01-01', '1985-03-31',convert=FALSE)

#> There are 90 data points, and 90 days.

# }