Import user daily data for EGRET analysis

Imports data from a user-supplied file, and converts it to a Daily data frame, appropriate for WRTDS calculations.

readUserDaily(filePath, fileName, hasHeader = TRUE, separator = ",",
  qUnit = 1, verbose = TRUE, interactive = NULL)

Arguments

filePath
character specifying the path to the file

fileName
character name of file to open

hasHeader
logical true if the first row of data is the column headers

separator
character character that separates data cells

qUnit
number 1 is cubic feet per second, 2 is cubic meters per second, 3 is 10^3 cubic feet per second, and 4 is 10^3 cubic meters per second

verbose
logical specifying whether or not to display progress message

interactive
logical deprecated. Use 'verbose' instead

Value

A data frame 'Daily' with the following columns:

Name	Type

Description		Date

Date	Date

Q	numeric	Discharge in m^3/s

Julian	integer

Number of days since Jan. 1, 1850		Month

integer	Month of the year [1-12]

Day	integer	Day of the year [1-366]

DecYear	numeric

Decimal year		MonthSeq

integer	Number of months since January 1, 1850

Qualifier	character	Qualifying code

i	integer

Index of days, starting with 1		LogQ

numeric	Natural logarithm of Q

Q7	numeric	7 day running average of Q

Q30	numeric

30 day running average of Q

Examples

filePath <- system.file("extdata", package="EGRET")
fileName <- "ChoptankRiverFlow.txt"
Daily <- readUserDaily(filePath,fileName,separator="\t")

#> The input discharge are assumed to be in cubic feet per second, if they are in cubic meters per second, then the call to readUserDaily should specify qUnit=2

#> There are 4383 data points, and 4383 days.