Example eList

Example data representing data from the Choptank River at Greensboro, MD, USGS data Data is a named list of the Daily, Sample, INFO dataframes, and the surface matrix.

Examples

head(Choptank_eList$Daily)
#>         Date        Q Julian Month Day  DecYear MonthSeq Qualifier i      LogQ
#> 1 1979-10-01 1.897229  47389    10 275 1979.750     1558         A 1 0.6403942
#> 2 1979-10-02 2.010496  47390    10 276 1979.753     1558         A 2 0.6983815
#> 3 1979-10-03 2.746734  47391    10 277 1979.755     1558         A 3 1.0104126
#> 4 1979-10-04 3.851091  47392    10 278 1979.758     1558         A 4 1.3483565
#> 5 1979-10-05 4.077626  47393    10 279 1979.761     1558         A 5 1.4055149
#> 6 1979-10-06 3.199804  47394    10 280 1979.764     1558         A 6 1.1630894
#>   Q7 Q30       yHat        SE   ConcDay  FluxDay    FNConc   FNFlux
#> 1 NA  NA -0.1645456 0.2214490 0.8745462 143.3561 0.9727565 114.7660
#> 2 NA  NA -0.1802420 0.2209740 0.8597670 149.3474 0.9828022 112.1125
#> 3 NA  NA -0.3044300 0.2170166 0.7641524 181.3470 0.9889741 109.7710
#> 4 NA  NA -0.4642075 0.2118471 0.6544177 217.7472 0.9940480 109.4190
#> 5 NA  NA -0.4872318 0.2110424 0.6383146 224.8826 1.0034610 108.3759
#> 6 NA  NA -0.3675283 0.2149243 0.7209948 199.3284 1.0084551 106.1974
head(Arkansas_eList$Daily)
#>            Date     Q Julian Month Day  DecYear MonthSeq    i     LogQ       Q7
#> 9132 1989-10-01 37600  51042    10 275 1989.749     1678 9132 10.53476 38071.43
#> 9133 1989-10-02 34400  51043    10 276 1989.752     1678 9133 10.44581 37071.43
#> 9134 1989-10-03 37200  51044    10 277 1989.755     1678 9134 10.52406 36400.00
#> 9135 1989-10-04 30200  51045    10 278 1989.758     1678 9135 10.31560 35571.43
#> 9136 1989-10-05 28100  51046    10 279 1989.760     1678 9136 10.24352 35185.71
#> 9137 1989-10-06 27100  51047    10 280 1989.763     1678 9137 10.20729 33642.86
#>           Q30 waterYear      yHat        SE ConcDay FluxDay FNConc FNFlux
#> 9132 51670.00      1990 -3.184867 0.7173803      NA      NA     NA     NA
#> 9133 51710.00      1990 -3.188835 0.7257386      NA      NA     NA     NA
#> 9134 51533.33      1990 -3.167081 0.7156364      NA      NA     NA     NA
#> 9135 51116.67      1990 -3.187290 0.7368824      NA      NA     NA     NA
#> 9136 51093.33      1990 -3.187121 0.7431933      NA      NA     NA     NA
#> 9137 51033.33      1990 -3.181600 0.7455803      NA      NA     NA     NA