Compute the 6 parameters needed to lay out the grid for the surfaces computed in estSurfaces

The code here is a repetition of the first part of the code for estSurfaces

Usage

surfaceIndex(Daily)

Arguments

Daily

data frame containing the daily values, default is Daily

Value

surfaceIndexParameters a numeric vector of length 6, defining the grid for the surfaces

Examples

eList <- Choptank_eList
Daily <- getDaily(eList)
surfaceIndex(Daily)
#> $bottomLogQ
#> [1] -4.66412
#> 
#> $stepLogQ
#> [1] 0.7862231
#> 
#> $nVectorLogQ
#> [1] 14
#> 
#> $bottomYear
#> [1] 1979
#> 
#> $stepYear
#> [1] 0.0625
#> 
#> $nVectorYear
#> [1] 529
#> 
#> $vectorYear
#>   [1] 1979.000 1979.062 1979.125 1979.188 1979.250 1979.312 1979.375 1979.438
#>   [9] 1979.500 1979.562 1979.625 1979.688 1979.750 1979.812 1979.875 1979.938
#>  [17] 1980.000 1980.062 1980.125 1980.188 1980.250 1980.312 1980.375 1980.438
#>  [25] 1980.500 1980.562 1980.625 1980.688 1980.750 1980.812 1980.875 1980.938
#>  [33] 1981.000 1981.062 1981.125 1981.188 1981.250 1981.312 1981.375 1981.438
#>  [41] 1981.500 1981.562 1981.625 1981.688 1981.750 1981.812 1981.875 1981.938
#>  [49] 1982.000 1982.062 1982.125 1982.188 1982.250 1982.312 1982.375 1982.438
#>  [57] 1982.500 1982.562 1982.625 1982.688 1982.750 1982.812 1982.875 1982.938
#>  [65] 1983.000 1983.062 1983.125 1983.188 1983.250 1983.312 1983.375 1983.438
#>  [73] 1983.500 1983.562 1983.625 1983.688 1983.750 1983.812 1983.875 1983.938
#>  [81] 1984.000 1984.062 1984.125 1984.188 1984.250 1984.312 1984.375 1984.438
#>  [89] 1984.500 1984.562 1984.625 1984.688 1984.750 1984.812 1984.875 1984.938
#>  [97] 1985.000 1985.062 1985.125 1985.188 1985.250 1985.312 1985.375 1985.438
#> [105] 1985.500 1985.562 1985.625 1985.688 1985.750 1985.812 1985.875 1985.938
#> [113] 1986.000 1986.062 1986.125 1986.188 1986.250 1986.312 1986.375 1986.438
#> [121] 1986.500 1986.562 1986.625 1986.688 1986.750 1986.812 1986.875 1986.938
#> [129] 1987.000 1987.062 1987.125 1987.188 1987.250 1987.312 1987.375 1987.438
#> [137] 1987.500 1987.562 1987.625 1987.688 1987.750 1987.812 1987.875 1987.938
#> [145] 1988.000 1988.062 1988.125 1988.188 1988.250 1988.312 1988.375 1988.438
#> [153] 1988.500 1988.562 1988.625 1988.688 1988.750 1988.812 1988.875 1988.938
#> [161] 1989.000 1989.062 1989.125 1989.188 1989.250 1989.312 1989.375 1989.438
#> [169] 1989.500 1989.562 1989.625 1989.688 1989.750 1989.812 1989.875 1989.938
#> [177] 1990.000 1990.062 1990.125 1990.188 1990.250 1990.312 1990.375 1990.438
#> [185] 1990.500 1990.562 1990.625 1990.688 1990.750 1990.812 1990.875 1990.938
#> [193] 1991.000 1991.062 1991.125 1991.188 1991.250 1991.312 1991.375 1991.438
#> [201] 1991.500 1991.562 1991.625 1991.688 1991.750 1991.812 1991.875 1991.938
#> [209] 1992.000 1992.062 1992.125 1992.188 1992.250 1992.312 1992.375 1992.438
#> [217] 1992.500 1992.562 1992.625 1992.688 1992.750 1992.812 1992.875 1992.938
#> [225] 1993.000 1993.062 1993.125 1993.188 1993.250 1993.312 1993.375 1993.438
#> [233] 1993.500 1993.562 1993.625 1993.688 1993.750 1993.812 1993.875 1993.938
#> [241] 1994.000 1994.062 1994.125 1994.188 1994.250 1994.312 1994.375 1994.438
#> [249] 1994.500 1994.562 1994.625 1994.688 1994.750 1994.812 1994.875 1994.938
#> [257] 1995.000 1995.062 1995.125 1995.188 1995.250 1995.312 1995.375 1995.438
#> [265] 1995.500 1995.562 1995.625 1995.688 1995.750 1995.812 1995.875 1995.938
#> [273] 1996.000 1996.062 1996.125 1996.188 1996.250 1996.312 1996.375 1996.438
#> [281] 1996.500 1996.562 1996.625 1996.688 1996.750 1996.812 1996.875 1996.938
#> [289] 1997.000 1997.062 1997.125 1997.188 1997.250 1997.312 1997.375 1997.438
#> [297] 1997.500 1997.562 1997.625 1997.688 1997.750 1997.812 1997.875 1997.938
#> [305] 1998.000 1998.062 1998.125 1998.188 1998.250 1998.312 1998.375 1998.438
#> [313] 1998.500 1998.562 1998.625 1998.688 1998.750 1998.812 1998.875 1998.938
#> [321] 1999.000 1999.062 1999.125 1999.188 1999.250 1999.312 1999.375 1999.438
#> [329] 1999.500 1999.562 1999.625 1999.688 1999.750 1999.812 1999.875 1999.938
#> [337] 2000.000 2000.062 2000.125 2000.188 2000.250 2000.312 2000.375 2000.438
#> [345] 2000.500 2000.562 2000.625 2000.688 2000.750 2000.812 2000.875 2000.938
#> [353] 2001.000 2001.062 2001.125 2001.188 2001.250 2001.312 2001.375 2001.438
#> [361] 2001.500 2001.562 2001.625 2001.688 2001.750 2001.812 2001.875 2001.938
#> [369] 2002.000 2002.062 2002.125 2002.188 2002.250 2002.312 2002.375 2002.438
#> [377] 2002.500 2002.562 2002.625 2002.688 2002.750 2002.812 2002.875 2002.938
#> [385] 2003.000 2003.062 2003.125 2003.188 2003.250 2003.312 2003.375 2003.438
#> [393] 2003.500 2003.562 2003.625 2003.688 2003.750 2003.812 2003.875 2003.938
#> [401] 2004.000 2004.062 2004.125 2004.188 2004.250 2004.312 2004.375 2004.438
#> [409] 2004.500 2004.562 2004.625 2004.688 2004.750 2004.812 2004.875 2004.938
#> [417] 2005.000 2005.062 2005.125 2005.188 2005.250 2005.312 2005.375 2005.438
#> [425] 2005.500 2005.562 2005.625 2005.688 2005.750 2005.812 2005.875 2005.938
#> [433] 2006.000 2006.062 2006.125 2006.188 2006.250 2006.312 2006.375 2006.438
#> [441] 2006.500 2006.562 2006.625 2006.688 2006.750 2006.812 2006.875 2006.938
#> [449] 2007.000 2007.062 2007.125 2007.188 2007.250 2007.312 2007.375 2007.438
#> [457] 2007.500 2007.562 2007.625 2007.688 2007.750 2007.812 2007.875 2007.938
#> [465] 2008.000 2008.062 2008.125 2008.188 2008.250 2008.312 2008.375 2008.438
#> [473] 2008.500 2008.562 2008.625 2008.688 2008.750 2008.812 2008.875 2008.938
#> [481] 2009.000 2009.062 2009.125 2009.188 2009.250 2009.312 2009.375 2009.438
#> [489] 2009.500 2009.562 2009.625 2009.688 2009.750 2009.812 2009.875 2009.938
#> [497] 2010.000 2010.062 2010.125 2010.188 2010.250 2010.312 2010.375 2010.438
#> [505] 2010.500 2010.562 2010.625 2010.688 2010.750 2010.812 2010.875 2010.938
#> [513] 2011.000 2011.062 2011.125 2011.188 2011.250 2011.312 2011.375 2011.438
#> [521] 2011.500 2011.562 2011.625 2011.688 2011.750 2011.812 2011.875 2011.938
#> [529] 2012.000
#> 
#> $vectorLogQ
#>  [1] -4.66412050 -3.87789739 -3.09167428 -2.30545117 -1.51922806 -0.73300495
#>  [7]  0.05321816  0.83944127  1.62566438  2.41188749  3.19811060  3.98433371
#> [13]  4.77055682  5.55677993
#>