Create a table of the changes in flow-normalized values between various points in time in the record

These tables describe trends in flow-normalized concentration and in flow-normalized flux. They are described as changes in real units or in percent and as slopes in real units per year or in percent per year. They are computed over pairs of time points. These time points can be user-defined or they can be set by the program to be the final year of the record and a set of years that are multiples of 5 years prior to that.

tableChange(eList, fluxUnit = 9, yearPoints = NA)

tableChangeSingle(eList, fluxUnit = 9, yearPoints = NA, flux = FALSE)

Arguments

eList
named list with at least the Daily and INFO dataframes

fluxUnit
object of fluxUnit class. printFluxUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name.

yearPoints
numeric vector listing the years for which the change or slope computations are made, they need to be in chronological order. For example yearPoints=c(1975,1985,1995,2005), default is NA (which allows the program to set yearPoints automatically)

flux
logical if TRUE results are returned in flux, if FALSE concentration. Default is set to FALSE.

Value

dataframe with Year1, Year2, change[mg/L], slope[mg/L], change[percent], slope[percent] columns. The data in each row is the change or slope calculated from Year1 to Year2

Examples

eList <- Choptank_eList
# Water Year:
# \donttest{
tableChange(eList, fluxUnit=6, yearPoints=c(2001,2005,2008,2009))

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>            Concentration trends
#>    time span       change     slope    change     slope
#>                      mg/L   mg/L/yr        %       %/yr
#> 
#>  2001  to  2005     0.073     0.018       5.7       1.4
#>  2001  to  2008      0.13     0.019        10       1.5
#>  2001  to  2009      0.15     0.019        12       1.5
#>  2005  to  2008      0.06      0.02       4.5       1.5
#>  2005  to  2009     0.079      0.02       5.9       1.5
#>  2008  to  2009     0.018     0.018       1.3       1.3
#> 
#> 
#>                  Flux Trends
#>    time span          change        slope       change        slope
#>                  10^3 tons/yr   10^3 tons/yr /yr      %         %/yr
#>  2001  to  2005        0.006       0.0015          3.9         0.97
#>  2001  to  2008         0.01       0.0014          6.5         0.93
#>  2001  to  2009       0.0098       0.0012          6.4         0.79
#>  2005  to  2008       0.0041       0.0014          2.5         0.85
#>  2005  to  2009       0.0038      0.00095          2.4         0.59
#>  2008  to  2009     -0.00028     -0.00028        -0.17        -0.17

tableChange(eList, fluxUnit=9)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>            Concentration trends
#>    time span       change     slope    change     slope
#>                      mg/L   mg/L/yr        %       %/yr
#> 
#>  1981  to  1986     0.039    0.0079       3.9      0.79
#>  1981  to  1991      0.14     0.014        14       1.4
#>  1981  to  1996      0.21     0.014        21       1.4
#>  1981  to  2001      0.27     0.013        27       1.3
#>  1981  to  2006      0.36     0.015        36       1.5
#>  1981  to  2011      0.46     0.015        46       1.5
#>  1986  to  1991       0.1     0.021        10         2
#>  1986  to  1996      0.18     0.018        17       1.7
#>  1986  to  2001      0.23     0.015        22       1.5
#>  1986  to  2006      0.32     0.016        31       1.6
#>  1986  to  2011      0.42     0.017        40       1.6
#>  1991  to  1996      0.07     0.014       6.2       1.2
#>  1991  to  2001      0.13     0.013        11       1.1
#>  1991  to  2006      0.22     0.015        19       1.3
#>  1991  to  2011      0.31     0.016        27       1.4
#>  1996  to  2001     0.055     0.011       4.5      0.91
#>  1996  to  2006      0.15     0.015        12       1.2
#>  1996  to  2011      0.24     0.016        20       1.3
#>  2001  to  2006     0.094     0.019       7.4       1.5
#>  2001  to  2011      0.19     0.019        15       1.5
#>  2006  to  2011     0.095     0.019         7       1.4
#> 
#> 
#>                  Flux Trends
#>    time span          change        slope       change        slope
#>                   10^6 kg/yr    10^6 kg/yr /yr      %         %/yr
#>  1981  to  1986        0.011       0.0022           10            2
#>  1981  to  1991        0.022       0.0022           20            2
#>  1981  to  1996        0.026       0.0018           24          1.6
#>  1981  to  2001        0.032       0.0016           29          1.5
#>  1981  to  2006        0.039       0.0016           36          1.4
#>  1981  to  2011         0.04       0.0013           37          1.2
#>  1986  to  1991        0.011       0.0022          9.4          1.9
#>  1986  to  1996        0.015       0.0015           13          1.3
#>  1986  to  2001        0.021       0.0014           18          1.2
#>  1986  to  2006        0.028       0.0014           23          1.2
#>  1986  to  2011        0.029       0.0012           24         0.97
#>  1991  to  1996       0.0042      0.00085          3.3         0.65
#>  1991  to  2001       0.0097      0.00097          7.5         0.75
#>  1991  to  2006        0.017       0.0011           13         0.86
#>  1991  to  2011        0.018      0.00089           14         0.68
#>  1996  to  2001       0.0055       0.0011          4.1         0.82
#>  1996  to  2006        0.013       0.0013          9.4         0.94
#>  1996  to  2011        0.014        9e-04           10         0.67
#>  2001  to  2006       0.0071       0.0014          5.1            1
#>  2001  to  2011       0.0081      0.00081          5.8         0.58
#>  2006  to  2011      0.00096      0.00019         0.65         0.13

tableChange(eList, fluxUnit=9, flowNormYear=c(2001:2006, 2008:2009))

#> Error in tableChange(eList, fluxUnit = 9, flowNormYear = c(2001:2006,     2008:2009)): unused argument (flowNormYear = c(2001:2006, 2008:2009))

# Winter:
eList <- setPA(eList, paStart=12,paLong=3)
tableChange(eList, fluxUnit=6,yearPoints=c(2001,2005,2008,2009))

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Dec Jan Feb 
#> 
#>            Concentration trends
#>    time span       change     slope    change     slope
#>                      mg/L   mg/L/yr        %       %/yr
#> 
#>  2001  to  2005       0.1     0.026       7.6       1.9
#>  2001  to  2008      0.18     0.026        13       1.9
#>  2001  to  2009      0.19     0.024        14       1.7
#>  2005  to  2008     0.076     0.025       5.1       1.7
#>  2005  to  2009     0.086     0.021       5.8       1.5
#>  2008  to  2009      0.01      0.01      0.65      0.65
#> 
#> 
#>                  Flux Trends
#>    time span          change        slope       change        slope
#>                  10^3 tons/yr   10^3 tons/yr /yr      %         %/yr
#>  2001  to  2005        0.012       0.0029          5.2          1.3
#>  2001  to  2008        0.019       0.0027          8.4          1.2
#>  2001  to  2009        0.018       0.0022          8.1            1
#>  2005  to  2008        0.007       0.0023            3            1
#>  2005  to  2009       0.0063       0.0016          2.7         0.67
#>  2008  to  2009     -0.00076     -0.00076        -0.32        -0.32

# Water Year:
#This returns concentration ASCII table in the console:
tableChangeSingle(eList, fluxUnit=6,yearPoints=c(2001,2005,2008,2009), flux=FALSE)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Dec Jan Feb 
#> 
#>            Concentration trends
#>    time span       change     slope    change     slope
#>                      mg/L   mg/L/yr        %       %/yr
#> 
#>  2001  to  2005       0.1     0.026       7.6       1.9
#>  2001  to  2008      0.18     0.026        13       1.9
#>  2001  to  2009      0.19     0.024        14       1.7
#>  2005  to  2008     0.076     0.025       5.1       1.7
#>  2005  to  2009     0.086     0.021       5.8       1.5
#>  2008  to  2009      0.01      0.01      0.65      0.65

#Returns a data frame:
change <- tableChangeSingle(eList, fluxUnit=6,yearPoints=c(2001,2005,2008,2009),
                  flowNormYears=c(2003:2004, 2006:2009), flux=FALSE)

#> Error in tableChangeSingle(eList, fluxUnit = 6, yearPoints = c(2001, 2005,     2008, 2009), flowNormYears = c(2003:2004, 2006:2009), flux = FALSE): unused argument (flowNormYears = c(2003:2004, 2006:2009))

#This returns flux values ASCII table in the console
df <- tableChangeSingle(eList, fluxUnit=6,yearPoints=c(2001,2005,2008,2009), flux=TRUE)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Dec Jan Feb 
#> 
#> 
#> 
#>                  Flux Trends
#>    time span          change        slope       change        slope
#>                 10^3 tons/yr  10^3 tons/yr/yr      %         %/yr
#> 
#>  2001  to  2005        0.012       0.0029          5.2          1.3
#>  2001  to  2008        0.019       0.0027          8.4          1.2
#>  2001  to  2009        0.018       0.0022          8.1            1
#>  2005  to  2008        0.007       0.0023            3            1
#>  2005  to  2009       0.0063       0.0016          2.7         0.67
#>  2008  to  2009     -0.00076     -0.00076        -0.32        -0.32

# Winter:
eList <- setPA(eList, paStart=12,paLong=3)
tableChangeSingle(eList, fluxUnit=6,yearPoints=c(2001,2005,2008,2009), flux=FALSE)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Dec Jan Feb 
#> 
#>            Concentration trends
#>    time span       change     slope    change     slope
#>                      mg/L   mg/L/yr        %       %/yr
#> 
#>  2001  to  2005       0.1     0.026       7.6       1.9
#>  2001  to  2008      0.18     0.026        13       1.9
#>  2001  to  2009      0.19     0.024        14       1.7
#>  2005  to  2008     0.076     0.025       5.1       1.7
#>  2005  to  2009     0.086     0.021       5.8       1.5
#>  2008  to  2009      0.01      0.01      0.65      0.65

# }