Table of annual results for discharge, concentration and flux

Produce an ASCII table showing: year, mean discharge, mean concentration, flow-normalized concentration, mean flux, and flow-normalized flux. Note that the flux and flow-normalized flux are rates and not a mass. As such a value for some period shorter than a full year could be larger than the value for a full year.

tableResults(eList, qUnit = 2, fluxUnit = 9, localDaily = NA)

Arguments

eList
named list with at least Daily and INFO dataframes

qUnit
object of qUnit class. printqUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name.

fluxUnit
object of fluxUnit class. printFluxUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name.

localDaily
data frame to override eList$Daily

Value

results dataframe, if returnDataFrame=TRUE

dataframe with year, discharge, concentration, flow-normalized concentration, flux, and flow-normalized concentration columns.

Examples

eList <- Choptank_eList
# Water Year:

tableResults(eList, fluxUnit = 1)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L             lbs/day   
#> 
#>    1980      4.25     0.949     1.003       697       642
#>    1981      2.22     1.035     0.999       407       654
#>    1982      3.05     1.036     0.993       594       663
#>    1983      4.99     1.007     0.993       802       674
#>    1984      5.72     0.990     1.002       964       688
#>    1985      1.52     1.057     1.017       295       702
#>    1986      2.63     1.062     1.038       545       720
#>    1987      3.37     1.079     1.062       690       739
#>    1988      1.87     1.120     1.085       399       756
#>    1989      5.61     1.055     1.105       989       768
#>    1990      4.01     1.115     1.125       814       779
#>    1991      2.75     1.172     1.143       592       787
#>    1992      2.19     1.203     1.159       489       794
#>    1993      3.73     1.215     1.173       788       797
#>    1994      5.48     1.144     1.187       986       801
#>    1995      2.41     1.266     1.201       560       806
#>    1996      6.24     1.134     1.213      1195       813
#>    1997      5.83     1.180     1.221      1137       818
#>    1998      4.88     1.236     1.229       962       824
#>    1999      2.90     1.277     1.238       555       831
#>    2000      4.72     1.213     1.253       982       840
#>    2001      4.88     1.251     1.268       999       846
#>    2002      1.24     1.321     1.285       291       853
#>    2003      8.64     1.140     1.303      1608       861
#>    2004      5.28     1.274     1.321      1106       870
#>    2005      3.81     1.360     1.341       871       879
#>    2006      3.59     1.382     1.362       850       889
#>    2007      4.28     1.408     1.382       961       897
#>    2008      2.56     1.477     1.401       608       902
#>    2009      3.68     1.409     1.419       801       900
#>    2010      7.19     1.323     1.438      1350       898
#>    2011      5.24     1.438     1.457       938       895

tableResults(eList, fluxUnit = 1, flowNormYears = c(1980:1995, 1997:2002, 2004:2011))

#> Error in tableResults(eList, fluxUnit = 1, flowNormYears = c(1980:1995,     1997:2002, 2004:2011)): unused argument (flowNormYears = c(1980:1995, 1997:2002, 2004:2011))

tableResults(eList, fluxUnit = 'kgDay', qUnit = 'cms')

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L              kg/day   
#> 
#>    1980      4.25     0.949     1.003       316       291
#>    1981      2.22     1.035     0.999       185       297
#>    1982      3.05     1.036     0.993       270       301
#>    1983      4.99     1.007     0.993       364       306
#>    1984      5.72     0.990     1.002       437       312
#>    1985      1.52     1.057     1.017       134       319
#>    1986      2.63     1.062     1.038       247       327
#>    1987      3.37     1.079     1.062       313       335
#>    1988      1.87     1.120     1.085       181       343
#>    1989      5.61     1.055     1.105       448       349
#>    1990      4.01     1.115     1.125       369       353
#>    1991      2.75     1.172     1.143       268       357
#>    1992      2.19     1.203     1.159       222       360
#>    1993      3.73     1.215     1.173       357       362
#>    1994      5.48     1.144     1.187       447       363
#>    1995      2.41     1.266     1.201       254       366
#>    1996      6.24     1.134     1.213       542       369
#>    1997      5.83     1.180     1.221       516       371
#>    1998      4.88     1.236     1.229       436       374
#>    1999      2.90     1.277     1.238       252       377
#>    2000      4.72     1.213     1.253       445       381
#>    2001      4.88     1.251     1.268       453       384
#>    2002      1.24     1.321     1.285       132       387
#>    2003      8.64     1.140     1.303       729       390
#>    2004      5.28     1.274     1.321       502       395
#>    2005      3.81     1.360     1.341       395       399
#>    2006      3.59     1.382     1.362       386       403
#>    2007      4.28     1.408     1.382       436       407
#>    2008      2.56     1.477     1.401       276       409
#>    2009      3.68     1.409     1.419       363       408
#>    2010      7.19     1.323     1.438       612       407
#>    2011      5.24     1.438     1.457       425       406

returnedTable <- tableResults(eList, fluxUnit = 1)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L             lbs/day   
#> 
#>    1980      4.25     0.949     1.003       697       642
#>    1981      2.22     1.035     0.999       407       654
#>    1982      3.05     1.036     0.993       594       663
#>    1983      4.99     1.007     0.993       802       674
#>    1984      5.72     0.990     1.002       964       688
#>    1985      1.52     1.057     1.017       295       702
#>    1986      2.63     1.062     1.038       545       720
#>    1987      3.37     1.079     1.062       690       739
#>    1988      1.87     1.120     1.085       399       756
#>    1989      5.61     1.055     1.105       989       768
#>    1990      4.01     1.115     1.125       814       779
#>    1991      2.75     1.172     1.143       592       787
#>    1992      2.19     1.203     1.159       489       794
#>    1993      3.73     1.215     1.173       788       797
#>    1994      5.48     1.144     1.187       986       801
#>    1995      2.41     1.266     1.201       560       806
#>    1996      6.24     1.134     1.213      1195       813
#>    1997      5.83     1.180     1.221      1137       818
#>    1998      4.88     1.236     1.229       962       824
#>    1999      2.90     1.277     1.238       555       831
#>    2000      4.72     1.213     1.253       982       840
#>    2001      4.88     1.251     1.268       999       846
#>    2002      1.24     1.321     1.285       291       853
#>    2003      8.64     1.140     1.303      1608       861
#>    2004      5.28     1.274     1.321      1106       870
#>    2005      3.81     1.360     1.341       871       879
#>    2006      3.59     1.382     1.362       850       889
#>    2007      4.28     1.408     1.382       961       897
#>    2008      2.56     1.477     1.401       608       902
#>    2009      3.68     1.409     1.419       801       900
#>    2010      7.19     1.323     1.438      1350       898
#>    2011      5.24     1.438     1.457       938       895

# Winter:
eList <- setPA(eList, paLong=3,paStart=12)
tableResults(eList, fluxUnit = 1)

#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Dec Jan Feb 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L             lbs/day   
#> 
#>    1980     4.220      1.10      1.11       847       939
#>    1981     1.960      1.20      1.13       444       972
#>    1982     5.057      1.16      1.15      1065       997
#>    1983     3.504      1.21      1.16       791      1019
#>    1984     8.437      1.12      1.18      1584      1044
#>    1985     2.249      1.24      1.21       533      1069
#>    1986     5.087      1.25      1.24      1148      1098
#>    1987     7.730      1.23      1.27      1634      1127
#>    1988     2.590      1.29      1.29       627      1151
#>    1989     3.327      1.36      1.31       845      1168
#>    1990     4.609      1.35      1.33      1164      1179
#>    1991     3.883      1.36      1.33       941      1182
#>    1992     2.303      1.40      1.33       613      1180
#>    1993     4.421      1.36      1.32      1133      1172
#>    1994     5.849      1.31      1.31      1250      1165
#>    1995     3.693      1.37      1.31       931      1161
#>    1996     7.151      1.29      1.31      1580      1162
#>    1997    11.786      1.18      1.32      2244      1165
#>    1998     9.308      1.27      1.32      1798      1172
#>    1999     2.526      1.32      1.33       659      1183
#>    2000     4.785      1.39      1.35      1204      1196
#>    2001     5.392      1.39      1.37      1324      1208
#>    2002     0.824      1.29      1.39       203      1220
#>    2003     9.213      1.33      1.41      1923      1234
#>    2004     8.653      1.35      1.44      1887      1252
#>    2005     4.196      1.54      1.47      1194      1271
#>    2006     5.843      1.51      1.51      1614      1291
#>    2007     5.417      1.55      1.53      1460      1305
#>    2008     2.436      1.62      1.55       735      1310
#>    2009     2.711      1.66      1.56       843      1305
#>    2010    13.779      1.26      1.57      2511      1299
#>    2011     3.369      1.66      1.57      1007      1295