Table of annual results for discharge, concentration and flux

Produce an ASCII table showing: year, mean discharge, mean concentration, flow-normalized concentration, mean flux, and flow-normalized flux. Note that the flux and flow-normalized flux are rates and not a mass. As such a value for some period shorter than a full year could be larger than the value for a full year.

Usage

tableResults(eList, qUnit = 2, fluxUnit = 9, localDaily = NA)

Arguments

eList

named list with at least Daily and INFO dataframes

qUnit

object of qUnit class. printqUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name.

fluxUnit

object of fluxUnit class. printFluxUnitCheatSheet, or numeric represented the short code, or character representing the descriptive name.

localDaily

data frame to override eList$Daily

Value

results dataframe, if returnDataFrame=TRUE

dataframe with year, discharge, concentration, flow-normalized concentration, flux, and flow-normalized concentration columns. If the eList was run through WRTDSKalman, an additional column generalized flux is included.

Details

Can also procude a table for any Period of Analysis (individual months or sequence of months) using setPA.

Examples

eList <- Choptank_eList
# Water Year:
# \donttest{
tableResults(eList, fluxUnit = 8)
#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L             10^3 kg/yr 
#> 
#>    1980      4.25     0.949     1.003     115.4       106
#>    1981      2.22     1.035     0.999      67.5       108
#>    1982      3.05     1.036     0.993      98.5       110
#>    1983      4.99     1.007     0.993     132.9       112
#>    1984      5.72     0.990     1.002     159.7       114
#>    1985      1.52     1.057     1.017      48.9       116
#>    1986      2.63     1.062     1.038      90.3       119
#>    1987      3.37     1.079     1.062     114.2       122
#>    1988      1.87     1.120     1.085      66.0       125
#>    1989      5.61     1.055     1.105     163.8       127
#>    1990      4.01     1.115     1.125     134.9       129
#>    1991      2.75     1.172     1.143      98.0       130
#>    1992      2.19     1.203     1.159      81.0       132
#>    1993      3.73     1.215     1.173     130.6       132
#>    1994      5.48     1.144     1.187     163.4       133
#>    1995      2.41     1.266     1.201      92.8       134
#>    1996      6.24     1.134     1.213     198.0       135
#>    1997      5.83     1.180     1.221     188.4       136
#>    1998      4.88     1.236     1.229     159.3       137
#>    1999      2.90     1.277     1.238      91.9       138
#>    2000      4.72     1.213     1.253     162.7       139
#>    2001      4.88     1.251     1.268     165.5       140
#>    2002      1.24     1.321     1.285      48.3       141
#>    2003      8.64     1.140     1.303     266.4       143
#>    2004      5.28     1.274     1.321     183.2       144
#>    2005      3.81     1.360     1.341     144.4       146
#>    2006      3.59     1.382     1.362     140.9       147
#>    2007      4.28     1.408     1.382     159.3       149
#>    2008      2.56     1.477     1.401     100.8       149
#>    2009      3.68     1.409     1.419     132.8       149
#>    2010      7.19     1.323     1.438     223.6       149
#>    2011      5.24     1.438     1.457     155.4       148
df <- tableResults(eList, fluxUnit = 1)
#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Water Year 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cms            mg/L             lbs/day   
#> 
#>    1980      4.25     0.949     1.003       697       642
#>    1981      2.22     1.035     0.999       407       654
#>    1982      3.05     1.036     0.993       594       663
#>    1983      4.99     1.007     0.993       802       674
#>    1984      5.72     0.990     1.002       964       688
#>    1985      1.52     1.057     1.017       295       702
#>    1986      2.63     1.062     1.038       545       720
#>    1987      3.37     1.079     1.062       690       739
#>    1988      1.87     1.120     1.085       399       756
#>    1989      5.61     1.055     1.105       989       768
#>    1990      4.01     1.115     1.125       814       779
#>    1991      2.75     1.172     1.143       592       787
#>    1992      2.19     1.203     1.159       489       794
#>    1993      3.73     1.215     1.173       788       797
#>    1994      5.48     1.144     1.187       986       801
#>    1995      2.41     1.266     1.201       560       806
#>    1996      6.24     1.134     1.213      1195       813
#>    1997      5.83     1.180     1.221      1137       818
#>    1998      4.88     1.236     1.229       962       824
#>    1999      2.90     1.277     1.238       555       831
#>    2000      4.72     1.213     1.253       982       840
#>    2001      4.88     1.251     1.268       999       846
#>    2002      1.24     1.321     1.285       291       853
#>    2003      8.64     1.140     1.303      1608       861
#>    2004      5.28     1.274     1.321      1106       870
#>    2005      3.81     1.360     1.341       871       879
#>    2006      3.59     1.382     1.362       850       889
#>    2007      4.28     1.408     1.382       961       897
#>    2008      2.56     1.477     1.401       608       902
#>    2009      3.68     1.409     1.419       801       900
#>    2010      7.19     1.323     1.438      1350       898
#>    2011      5.24     1.438     1.457       938       895
df
#>    Year Discharge [cms] Conc [mg/L] FN Conc [mg/L] Flux [lbs/day]
#> 1  1980            4.25       0.949          1.003            697
#> 2  1981            2.22       1.035          0.999            407
#> 3  1982            3.05       1.036          0.993            594
#> 4  1983            4.99       1.007          0.993            802
#> 5  1984            5.72       0.990          1.002            964
#> 6  1985            1.52       1.057          1.017            295
#> 7  1986            2.63       1.062          1.038            545
#> 8  1987            3.37       1.079          1.062            690
#> 9  1988            1.87       1.120          1.085            399
#> 10 1989            5.61       1.055          1.105            989
#> 11 1990            4.01       1.115          1.125            814
#> 12 1991            2.75       1.172          1.143            592
#> 13 1992            2.19       1.203          1.159            489
#> 14 1993            3.73       1.215          1.173            788
#> 15 1994            5.48       1.144          1.187            986
#> 16 1995            2.41       1.266          1.201            560
#> 17 1996            6.24       1.134          1.213           1195
#> 18 1997            5.83       1.180          1.221           1137
#> 19 1998            4.88       1.236          1.229            962
#> 20 1999            2.90       1.277          1.238            555
#> 21 2000            4.72       1.213          1.253            982
#> 22 2001            4.88       1.251          1.268            999
#> 23 2002            1.24       1.321          1.285            291
#> 24 2003            8.64       1.140          1.303           1608
#> 25 2004            5.28       1.274          1.321           1106
#> 26 2005            3.81       1.360          1.341            871
#> 27 2006            3.59       1.382          1.362            850
#> 28 2007            4.28       1.408          1.382            961
#> 29 2008            2.56       1.477          1.401            608
#> 30 2009            3.68       1.409          1.419            801
#> 31 2010            7.19       1.323          1.438           1350
#> 32 2011            5.24       1.438          1.457            938
#>    FN Flux [lbs/day]
#> 1                642
#> 2                654
#> 3                663
#> 4                674
#> 5                688
#> 6                702
#> 7                720
#> 8                739
#> 9                756
#> 10               768
#> 11               779
#> 12               787
#> 13               794
#> 14               797
#> 15               801
#> 16               806
#> 17               813
#> 18               818
#> 19               824
#> 20               831
#> 21               840
#> 22               846
#> 23               853
#> 24               861
#> 25               870
#> 26               879
#> 27               889
#> 28               897
#> 29               902
#> 30               900
#> 31               898
#> 32               895
# Spring:
eList <- setPA(eList, paStart = 3, paLong = 3)
tableResults(eList, fluxUnit = 1, qUnit = "cfs")
#> 
#>    Choptank River 
#>    Inorganic nitrogen (nitrate and nitrite)
#>    Season Consisting of Mar Apr May 
#> 
#>    Year   Discharge    Conc    FN_Conc     Flux    FN_Flux
#>              cfs            mg/L             lbs/day   
#> 
#>    1980     236.6     0.921     0.922      1104      1006
#>    1981     132.1     0.957     0.936       660      1028
#>    1982     165.2     0.972     0.941       873      1038
#>    1983     430.5     0.879     0.947      1804      1051
#>    1984     365.4     0.917     0.958      1604      1068
#>    1985      64.8     1.007     0.973       357      1086
#>    1986     115.7     1.030     0.993       662      1111
#>    1987     167.8     1.036     1.016       927      1138
#>    1988     129.1     1.065     1.038       724      1163
#>    1989     384.7     0.992     1.057      1886      1183
#>    1990     216.6     1.074     1.073      1142      1197
#>    1991     154.1     1.119     1.086       926      1204
#>    1992     141.2     1.134     1.096       856      1208
#>    1993     310.8     1.052     1.105      1634      1210
#>    1994     427.2     1.032     1.113      1962      1214
#>    1995     138.2     1.168     1.121       834      1221
#>    1996     303.3     1.078     1.130      1641      1230
#>    1997     241.2     1.124     1.137      1364      1239
#>    1998     272.6     1.113     1.145      1476      1247
#>    1999     151.9     1.203     1.154       955      1253
#>    2000     275.0     1.139     1.164      1465      1259
#>    2001     240.4     1.166     1.172      1356      1264
#>    2002      94.4     1.239     1.181       610      1269
#>    2003     356.6     1.082     1.192      1903      1276
#>    2004     207.7     1.229     1.205      1186      1286
#>    2005     283.1     1.181     1.219      1561      1298
#>    2006      80.4     1.362     1.234       586      1308
#>    2007     265.5     1.252     1.246      1415      1315
#>    2008     213.7     1.268     1.255      1267      1314
#>    2009     159.5     1.302     1.259      1027      1306
#>    2010     289.7     1.202     1.255      1534      1291
#>    2011     193.9     1.243     1.241      1218      1267
# }